requirements.txt
flask
ultralytics
torch
torchvision
opencv-python-headless
pandas
numpy
pillow
werkzeug